backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
opencv-python-headless>=4.8
ultralytics>=8.0
easyocr>=1.7
numpy>=1.24
roboflow>=1.1  # only needed for training seatbelt.pt via ai/train_seatbelt.py
